refactor: Improve file structure by features + refactor story page custom hooks
