+++ src/features/openai/domain/openai.ts
@@ -1,5 +1,10 @@
 import OpenAI from "openai";
 
+export enum OpenAIModelsEnum {
+  ImageInterpreter = "gpt-4-vision-preview",
+  Language = "gpt-4",
+}
+
 export const openAI = new OpenAI({
   apiKey: process.env.OPENAI_API_KEY,
 });
