+++ src/features/story/presentation/components/LevelEllipse.tsx
