+++ src/features/story/presentation/components/StoryBox.tsx
