+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
-import ContentContainer from "@/components/ContentContainer";
-import { Button } from "@/components/Button";
+import ContentContainer from "@/shared/components/organisms/ContentContainer";
+import { Button } from "@/shared/components/molecules";
 import Link from "next/link";
 import LogoSVG from "@/assets/images/Logo.svg";
