+++ src/shared/components/atoms/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Loading";
