+++ src/shared/components/molecules/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Button";
+export * from "./RoundedButton";
